+++ NEWS.md
@@ -1,3 +1,10 @@
+# LMMsolver 1.0.11
+
+-   New function `mLogLik()` for the calculations of the log-likelihood and first derivatives as function of precision parameters `theta`. 
+-   A new argument `deriv` added to `predict()` to calculate the first derivatives for `spl1D()` functions.
+-   Two examples in vignette updated with predictions of derivatives and corresponding standard errors.
+-   bug fixed for `theta` argument of `LMMsolve()`.
+
 # LMMsolver 1.0.10
 
 -   Cyclic B-splines models added for `spl1D()` and `spl2D()` functions.